Remove hard-coded ITRF ranges for ATM1B data
We expect every qfit file to contain a header with the ITRF given within. The
code responsible for reading headers originally failed to read some qfit headers
du to different lengths of the header in bytes. Now that code is fixed and we
should always get the source ITRF from the data itself, not a hard-coded list.