WIP refactor function that creates iceflow parquet
Idea is to make the iceflow data dir the center of the workflow. We download
data to it, and we read data from it. The iceflow parquet file facilitates
analysis of elevation data for the data within the data dir.

A user could have multiple data dirs for different areas of interest.