More flexible bounding box model
This is much more flexible, but classes that inherit BoundingBox as a field will
still raise a validation error if BoundingBox is not initialized with a dict or
model instance (e.g. `DatasetSearchParameters`). This is the most likely point
that users would define and use a bounding box. Ideally, a user could pass in a
list representing a BBOX at the `DatasetSearchParameters` level