Fixup issue with `.ipynb` extension
Sphinx/myst_nb seems to have problems with including the extension. I get xref
missing warnings when they are present. It also causes relative links with
`.ipynb` to not render at all in the docs.

This represents a trade-off: the docs will have nicely rendered links to the
other notebooks. For users who download the files themselves and execute them as
notebooks locally, the links will not work. However, it should be clear enough
what they refer to.