Move all notebooks into `doc/notebooks`
Eliminates confusion around notebooks "living" in two separate
locations. Although the symlinks did work, it was still a bit
confusing. Importantly, clicking on the "view this page" link at the top that
takes the user to GitHub directs users to the symlink, which might be
confusing. Users should be able to click on that button to directly go to and
download the notebook and run it themselves.

This also makes it clearer that the notebooks are part of the documentation.

It still may be worth considering letting `myst_nb` execute the notebooks
at build time, but I worry about very long execution times for the icepyx
notebook in particular. It might be better to have the developer execute the
cells manually. Also, it would require our CI pipeline to have earthdata login
creds to download the necessary data.